requirements.txt
Flask>=2.1.2
google-cloud-bigquery>=3.1.0
google-cloud-storage>=1.31.0
gunicorn>=20.1.0
pandas>=1.4.2
json5>=0.9.6
